+++ src/icons/clothing-fashion/index.ts
@@ -1,0 +1,3 @@
+import DiamondRing from '@/src/icons/clothing-fashion/DiamondRing';
+
+export { DiamondRing };