+++ src/icons/devices-hardware/index.ts
@@ -1,5 +1,5 @@
-import Airpodes from '@/src/icons/devices-hardware/Airpodes';
+import Airpods from '@/src/icons/devices-hardware/Airpods';
 import Devices from '@/src/icons/devices-hardware/Devices';
 import USBACable from '@/src/icons/devices-hardware/USBACable';
 
-export { Airpodes, Devices, USBACable };
+export { Airpods, Devices, USBACable };