+++ src/icons/education/index.ts
@@ -1,0 +1,3 @@
+import OpenBook from '@/src/icons/education/OpenBook';
+
+export { OpenBook };