+++ src/icons/food-beverage/index.ts
@@ -1,0 +1,3 @@
+import EatenApple from '@/src/icons/food-beverage/EatenApple';
+
+export { EatenApple };